Updates download link in installation docs
This reverts the previous commit, which instructed the user to download the latest version, where the expected behaviour was downloading the latest *release*.
There does not seem to exist an automatic (i.e. not explicit) way of downloading the latest release. For now the link needs to be updated at each release.