Bug fix in prior transform with Ultranest
In contrast to Dynasty and Pymultinest, Ultranest requires the prior transform to
internally copy the incoming cube (As can be seen in the Ultranest demos,
e.g. https://johannesbuchner.github.io/UltraNest/example-sine-line.html.)
All samplers still need the untransformed samples to correctly estimate the pior mass, however,
both Dynasty and Pymultinest catch a potential inplace modification internally.
In contrast, Ultranest only checks if the untransformed sample has been transformed out of the unit cube, which catches
most prior transforms which do not copy internally.
This was not the case for IMAGINE, however, as we were forcing a unit cube even for the transformed samples.
This means that all previous IMAGINE runs which were using Ultranest and NOT a flat prior are probably wrong.