Includes MPI barrier in the end of load_pipeline
This change avoids synchronization errors. These could appear, for example, if one submits a script to a cluster which loads a given pipeline and runs it. If there was no barrier and one of the MPI processes is slower than the rank=0, the slow process would still be loading while the master was already starting the run and saving it. This situation would lead the worker process to read a truncated (bogus) pipeline pickle file. Similar bugs could affect a wrapped sampler or model. With the barrier, everyone is at the same point after loading, avoiding this type of problem.